Update baselevel_learning function for hard-coded activations
According to the ACT-R manual, hard-coded activations only apply if base-level learning is off.

"If :bll is nil then the setting of :ol does not matter and the base-level is a constant value determined by the :blc parameter or specific user settings for the chunk." (pg. 291)

Also, if bll is off, the activation should default to 0.0 (:blc defaults to 0.0 in ACT-R).